requirements.txt
streamlit
openai
google-generativeai
pandas
matplotlib
nltk
seaborn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread-dataframe
python-dotenv
plotly
numpy
pyyaml
toml
